pyproject.toml
[project]
name = "test-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-adk>=1.2.1",
    "google-search-results>=2.4.2",
    "litellm>=1.72.2",
    "pydantic>=2.11.5",
]
